+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - (04.02.2024)
+## Changes
+- add support for Light - On/Off, Brightness, Color Temperature, Hue, Saturation
+- add possibility to disable/enable name prefix for relays, ligts, sensors
+- config schema updated
+- cleanu
+
 ## [0.8.0] - (03.02.2024)
 ## Changes
 - add support for Ambient Light and Carbon Dioxyde Sensors